requirements.txt
dask==2021.10.0
distributed==2021.10.0
django==3.2.8
django-core-headers==3.10.0
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
markdown==3.3.4
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
scipy==1.7.1
sklearn==1.0.1
uwsgi==2.0.20